+++ apps/prime-ng-playground/src/app/app.routes.ts
@@ -1,0 +1,13 @@
+import { Route } from '@angular/router';
+
+export const appRoutes: Route[] = [
+  {
+    path: 'home',
+    loadComponent: () => import('../pages/home/home.component').then((m) => m.HomeComponent),
+  },
+  {
+    path: '**',
+    pathMatch: 'full',
+    redirectTo: 'home',
+  },
+];