+++ libs/core/src/index.ts
@@ -12,6 +12,7 @@ export * from './lib/services/interpolation.service';
 export * from './lib/services/remote-resource/remote-resource.interface';
 export * from './lib/services/remote-resource/remote-resource.service';
 export * from './lib/services/state-store.service';
+export * from './lib/services/templates/base-template.service';
 export * from './lib/services/templates/layout-template.service';
 export * from './lib/services/templates/layout-template-interfaces';
 export * from './lib/services/templates/remote-resource-template.service';